chore: separate `gen_inclusion_verifier` from `gen_inclusion_proof`
